requirements.txt
feedparser
gTTS
moviepy
pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
